refactor(ai): split SwarmHeartbeatService into class + entry-point wrapper (#11058)
Closes the entry-point-only invariant gap surfaced during PR #11054 (#11049):
SwarmHeartbeatService was a hybrid file (class + self-invoke at the bottom),
the only daemon-tier class still importing Neo directly post-#11054. This PR
applies the canonical Orchestrator class+wrapper pattern (#11041 / #11044 /
#11049 precedent):

- `ai/daemons/SwarmHeartbeatService.mjs`: class-only file. Removed Neo +
  core/_export + InstanceManager imports + self-invoke `if (isMain)` block at
  the bottom. Now matches sibling pattern (TaskStateService /
  ProcessSupervisorService / SummarizationCoordinatorService / DreamService).

- `ai/scripts/swarm-heartbeat-daemon.mjs` (NEW): entry-point wrapper. Owns Neo
  namespace bootstrap + SIGTERM / SIGINT clean-stop signal handlers +
  `SwarmHeartbeatService.start()` invocation. Mirrors orchestrator-daemon.mjs
  shape.

- `test/playwright/unit/ai/daemons/SwarmHeartbeatService.spec.mjs`: added
  `import Neo + import * as core` at top to bootstrap the Neo namespace before
  the dynamic SwarmHeartbeatService import. Matches the test-spec-as-entry-point
  pattern from sibling specs (TaskStateService.spec / ProcessSupervisorService.spec
  / SummarizationCoordinatorService.spec / Orchestrator.spec post-#11049).

- `learn/agentos/wake-substrate/PersistentProcessManagement.md`: updated
  invocation paths from `node ai/daemons/SwarmHeartbeatService.mjs` →
  `node ai/scripts/swarm-heartbeat-daemon.mjs` (lines 35, 46, 188 + descriptive
  references). The class file itself is still the daemon source (line 231).

- `learn/agentos/wake-substrate/com.neomjs.swarm-heartbeat.plist.template`:
  updated launchd ProgramArguments path + Target comment to point at the
  wrapper.

Empirical:
- 14/14 SwarmHeartbeatService.spec tests pass
- 52/52 MailboxService.spec tests pass (no collateral damage)
- node --check passes for both class file + new wrapper
- Net diff: -49 / +31 lines (net reduction; substrate accretion defense
  satisfied — the split removes redundant comment-prelude + self-invoke
  block, replacing with smaller wrapper file)

Operator action required post-merge:
- launchd / systemd installations using the old direct-class-file path will
  need to update their plist / .service file to target
  `ai/scripts/swarm-heartbeat-daemon.mjs`. The .plist.template already updated;
  re-running §3a sed substitution against the new template captures it.

Self-Identification: @neo-opus-4-7 (Claude Opus 4.7, Claude Code)
Origin Session ID: c2912891-b459-4a03-b2af-154d5e264df1
